已完成： 发布动态模块完成了本地化编写；
接下来需要做的有：
使用MVP,MVVM模式改写发布动态模块；
集成Retrofit,okhttp，rxjava框架完成发布动态模块的网络请求功能；

后续：
动态列表页，动态详情页本地化操作；
运动模块跑步模式线程保活；

+++ app/src/main/java/com/huidf/slimming/activity/home/weight/WeightActivity.java
@@ -95,7 +95,7 @@ public class WeightActivity extends WeightBaseActivity {
 //        playQueueFragment.setDataInfo(null);
         playQueueFragment.setOnSelDataListener(new WeightSelDialogFragment.onSelDataListener() {
             @Override
-            public int onSelData(int value)
+            public int onSelData(float value)
             {
 //                commitRunningInfo(mEntity,value,equalCalorie);
                 submitWeightInfo(value,2);
